+++ scripts/generate-pyproject.py
@@ -406,7 +406,7 @@ write_project_files(
     PROJECTION_PATH / "winrt-runtime",
     "winrt.system",
     "_winrt",
-    ["_winrt.cpp", "_winrt_array.cpp", "runtime.cpp"],
+    ["_winrt.cpp", "_winrt_array.cpp", "_winrt_box.cpp", "runtime.cpp"],
 )
 
 for package_path in chain(